server/requirements.txt
openenv-core[core]>=0.3.0
numpy>=1.26.0
pandas>=2.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
openai>=1.30.0
setuptools>=70.0.0